Update https://w3c.github.io/mediacapture-record/#dom-blobeventinit-timecode.
The spec text was unclear. From offline discussion with the original spec author
it's believed that the intent of the spec language was meant to convey
that the first produced timecode would contain an absolute timestamp. However,
it's not clear what usecase is unlocked by that, as the user can construct
their own absolute timestamp with performance.now() on event reception. This
ignores the fact that the event handler can execute delayed, but since it's
not clear what problem this fixes, this PR simply makes the timestamp sequence
start with zero instead.

Fixes #222.